CHPIstanbul/backend/requirements.txt
# CHP İstanbul CRM — Django API
Django>=5.1,<6
djangorestframework>=3.15
django-cors-headers>=4.6
Pillow>=10
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-safedelete>=1.4.0
django-simple-history>=3.10.0
gunicorn>=23.0
whitenoise>=6.8
django-storages>=1.14
boto3>=1.35
